Extract modelstub for testing
