+++ src/test/java/seedu/address/logic/commands/add/AddResumeCommandTest.java
@@ -15,9 +15,10 @@ import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 
 import seedu.address.model.item.Resume;
+import seedu.address.testutil.ModelStub;
 import seedu.address.testutil.ResumeBuilder;
 
-public class AddResumeCommandTest extends AddCommandTest {
+public class AddResumeCommandTest {
     @Test
     public void constructor_nullResume_throwsNullPointerException() {
         assertThrows(NullPointerException.class, () -> new AddResumeCommand(null));
